Attempt to make the list of Docker versions a bit more manageable.
