+++ nodeup/pkg/model/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "architecture.go",
         "cloudconfig.go",
         "context.go",
         "convenience.go",
