+++ nodeup/pkg/model/resources/BUILD.bazel
@@ -2,7 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["docker.go"],
+    srcs = [
+        "architecture.go",
+        "docker.go",
+        "docker_centos.go",
+        "docker_debian.go",
+        "docker_ubuntu.go",
+    ],
     importpath = "k8s.io/kops/nodeup/pkg/model/resources",
     visibility = ["//visibility:public"],
+    deps = ["//nodeup/pkg/distros:go_default_library"],
 )
